% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_guide}
\alias{export_guide}
\alias{export_guide_ui}
\alias{export_guide_server}
\alias{export_guide_app}
\title{Shiny modules providing the Export Guide}
\usage{
export_guide_ui(id)

export_guide_server(id, registry_name)

export_guide_app()
}
\arguments{
\item{id}{Character string module ID}

\item{registry_name}{Character string registry name key}
}
\value{
Functions ui and server representing the (module) app
}
\description{
Shiny modules providing the Export Guide
}
\examples{
ui <- shiny::fluidPage(
  export_guide_ui("guide")
)

server <- function(input, output, session) {
  export_guide_server("guide", "test")
}

if (interactive()) {
  shiny::shinyApp(ui, server)
}
}
